move and refactor tag and tagrel callback functions
