+++ packages/lesswrong/server/callbacks/votingCallbacks.ts
@@ -17,10 +17,12 @@ import { postGetPageUrl } from '../../lib/collections/posts/helpers';
 import { createManifoldMarket } from '../../lib/collections/posts/annualReviewMarkets';
 import { RECEIVED_SENIOR_DOWNVOTES_ALERT } from '../../lib/collections/moderatorActions/schema';
 import { revokeUserAFKarmaForCancelledVote, grantUserAFKarmaForVote } from './alignment-forum/callbacks';
-import { recomputeContributorScoresFor, voteUpdatePostDenormalizedTags } from '../tagging/tagCallbacks';
 import { updateModerateOwnPersonal, updateTrustedStatus } from '../users/moderationUtils';
 import { captureException } from '@sentry/core';
 import { tagGetUrl } from '@/lib/collections/tags/helpers';
+import { updatePostDenormalizedTags } from '../tagging/helpers';
+import { updateDenormalizedContributorsList } from '../utils/contributorsUtil';
+import { MultiDocuments } from '@/lib/collections/multiDocuments/collection';
 
 async function increaseMaxBaseScore({newDocument, vote}: VoteDocTuple) {
   if (vote.collectionName === "Posts") {
@@ -47,6 +49,33 @@ async function increaseMaxBaseScore({newDocument, vote}: VoteDocTuple) {
       }
       await Posts.rawUpdateOne({_id: post._id}, {$set: {maxBaseScore: post.baseScore, ...thresholdTimestamp}})
     }
+  }
+}
+
+function voteUpdatePostDenormalizedTags({newDocument}: {newDocument: VoteableType}) {
+  let postId: string;
+  if ("postId" in newDocument) { // is a tagRel
+    // Applying human knowledge here
+    postId = (newDocument as DbTagRel)["postId"];
+  } else if ("tagRelevance" in newDocument) { // is a post
+    postId = newDocument["_id"];
+  } else {
+    return;
+  }
+  void updatePostDenormalizedTags(postId);
+}
+
+export async function recomputeContributorScoresFor(votedRevision: DbRevision) {
+  if (votedRevision.collectionName !== "Tags" && votedRevision.collectionName !== "MultiDocuments") return;
+
+  if (votedRevision.collectionName === "Tags") {
+    const tag = await Tags.findOne({_id: votedRevision.documentId});
+    if (!tag) return;
+    await updateDenormalizedContributorsList({ document: tag, collectionName: 'Tags', fieldName: 'description' });
+  } else if (votedRevision.collectionName === "MultiDocuments") {
+    const multiDocument = await MultiDocuments.findOne({_id: votedRevision.documentId});
+    if (!multiDocument) return;
+    await updateDenormalizedContributorsList({ document: multiDocument, collectionName: 'MultiDocuments', fieldName: 'contents' });
   }
 }
